pyproject.toml
[project]
name = "multi-agent-research"
version = "0.1.0"
description = "Simple multi-agent research system using Claude Agents SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.51",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
    # Vendored skill resources — keep as-shipped.
    ".claude/skills",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PL", "RUF"]
ignore = [
    "PLR0913",  # too-many-arguments — factories legitimately have many kwargs
    "PLR2004",  # magic-value comparison — pragmatic in tests and small scripts
    "PLC0415",  # nested imports — used intentionally for module-import smoke tests
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]
"scripts/*" = ["T201"]  # print() is the whole point of these scripts
